Exact function imports
Function references to imported functions must normally be inexact to
preserve soundness in the case where the dynamic type of the imported
function is a subtype of the import type. However, this difference
between imported and declared functions harms modularity by making it
possible for an instruction sequence to become invalid when a function
it references is replaced with an import.

To fix the modularity problem, introduce exact function imports that can
be referenced exactly, just as defined functions can.

Describe the problem and solution in the overview, and also update the
interpreter to make references to normal imported functions inexact and
to support exact function imports.

Fixes #57.