Fix validation rules for descriptor types (#62)
Strengthen the rules to disallow a situation where a subtype of a
descriptor type describes a type that is unrelated to the type described
by the descriptor type. This situation was previously allowed, but
allowed for inexact descriptor casts to unsoundly cast from the the type
described by the descriptor subtype to the unrelated type described by
the descriptor supertype.

Disallow this situation from occuring at validation time to ensure that
inexact descriptor casts are sound.

This fix has already been implemented in Binaryen with no loss of
optimization power on a real-world code base.

Fixes #58.